+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixed the `--no-dev` option in `install` not working properly.
 - Fixed prereleases being selected even if another constraint conflicted with them.
 - Fixed an error when installing current package in development mode if the generated `setup.py` had special characters.
+- Fixed an error in `install` for applications not following a known structure.
 
 
 ## [0.12.2] - 2018-10-17